requirements.txt
selenium==3.141.0
Pillow==7.1.2
PySimpleGUI==4.19.0
instagram-scraper==1.9.0
PyAutoIt==0.6.0
